chore: add Search form to page in FE

+++ backend/src/main/java/com/greenfoxacademy/backend/repositories/PetRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * Repository to manage Pet entities.
  */
 public interface PetRepository extends JpaRepository<Pet, Long> {
-  List<Pet> findAllByOwnerId(Integer ownerId);
+  List<Pet> findAllByOwnerId(Long ownerId);
 }
